Added output file for Sub-task 2 of SIMMC 2.0
We developed Bi-Encoder and Poly-Encoder based models (Humeau et al., 2020). We regard the objects in the corresponding scenes as the candidate objects. We regard the previous system transcript and current user transcript as the context. And for each candidate object, we compute the dot-product between it and the context. If Sub-Task 1 returns a binary label (0 or 1) for the current turn, Sub-Task 2 will return top-K (K=2 as default) object IDs as the results.