Added output files for Sub-task 4 of SIMMC 2.0
For Sub-Task 1,3,4, we developed an end-to-end encoder-decoder model based on BART (Lewis et al., 2020) for generating outputs of the tasks (Sub-Task #1, Sub-Task #3 , Sub-Task #4 Response) in a single string, called joint learning model, and another variant for generating outputs of the Sub-Task #4 Retrieval, called retrieval model. The retrieval model trains on Sub-Task 4 only. The two models are trained and evaluated separately.