Added output file for sub-tasks 2,3
For Sub-Task 1,3,4, we developed an end-to-end encoder-decoder model based on BART (Lewis et al., 2020) for generating outputs of the tasks (Sub-Task #1, Sub-Task #3 , Sub-Task #4 Response) in a single string, called joint learning model, and another variant for generating outputs of the Sub-Task #4 Retrieval, called retrieval model. The retrieval model trains on Sub-Task 4 only. The two models are trained and evaluated separately. 

For Sub-Task 2, we developed Bi-Encoder and Poly-Encoder based models (Humeau et al., 2020). We regard the objects in the corresponding scenes as the candidate objects. We regard the previous system transcript and current user transcript as the context. And for each candidate object, we compute the dot-product between it and the context. If Sub-Task 1 returns a binary label (0 or 1) for the current turn, Sub-Task 2 will return top-K (K=2 as default) object IDs as the results.